Fixed the prometheus configuration issue
The regex pattern contains $1 and $2 to represent the matched
label part but using bash, $1 and $2 get replaced as empty string
when not set which produces wrong prometheus configuration file.
This patch fixes that issue by setting up a variable so the end
result of the prometheus configuration will be correct.

Signed-off-by: Tong Li <litong01@us.ibm.com>